setup.py: Use pkg-config to identify libcamera
libcamera can be installed by a distribution or from source and
the paths to include should be obtained by the pkg-config tool

Provide a helper to obtain those arguments, and use them when
creating the pylibcamera Extension.

This brings in camera_base as an additional library linkage
which was previously missed, and ensures consistent links
to the required include dirs.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>